api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas>=1.5.0
pydantic>=2.0.0
python-multipart==0.0.6
pymysql>=1.1.0
sqlalchemy>=2.0.0
cryptography>=41.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
